requirements.txt
gradio>=4.0,<5.0
z3-solver>=4.12
matplotlib>=3.7
pytest>=7.0
PyYAML>=6.0
jsonschema>=4.0
psutil>=5.9
pynvml>=11.0